From: BALATON Zoltan <balaton@eik.bme.hu> Date: Sun, 14 Jun 2020 20:03:06 +0200 Subject: [PATCH 2/2] mac_oldworld: Add SPD data to cover RAM
OpenBIOS gets RAM size via fw_cfg but the original board firmware
detects RAM using SPD data so generate and add SDP EEPROMs to cover as
much RAM as possible to describe with SPD (this may be less than the
specified ram_size due to SDRAM size constraints).

This patch is more complex as it should be which I intend to fix once
agreement can be made on how to get back the necessary functionality
removed by previous patches. See in this thread:
https://lists.nongnu.org/archive/html/qemu-devel/2020-06/msg08710.html